Initial commit
because: Passenger was rejecting requests with 503 errors
when it's pool is full. It was misconfigured artificially low,
but what if it wasn't too low? How could we instruct the load
balancer to check if it even should receive traffic? Putting more
load on the Passenger server itself isn't a great answer. Separate
health checks can be made later about the server.

this commit: Builds an initial application that
can return if the Passenger server can accept
more traffic into it's queue.

This was coded quickly, and the Rust can be made smoother and
more idiomatic. The tests are also not complete, but offer
and initial structure for running tests.